third_party/blink/web_tests/fast/backgrounds/background-repeat-computed-style-expected.txt

This div should have a horizontal repeating background (background-repeat:repeat-x)
This div should have a vertical repeating background (background-repeat:repeat-y)
This div should have a horizontal repeating background (background-repeat-y:no-repeat)
This div should have a vertical repeating background (background-repeat-x:no-repeat)
This div should have vertical and horizontal repeating background ()
This div should have vertical and horizontal repeating background (background-repeat-x:repeat; background-repeat-y:repeat)
This div should have a horizontal repeating background (background-repeat: no-repeat; background-repeat-x:repeat)
This div should have a vertical repeating background (background-repeat: no-repeat; background-repeat-y:repeat)
This div should have a vertical and horizontal repeating background (background-repeat: repeat;)
This div should have vertical and horizontal repeating background (background-repeat-x:repeat; background-repeat-y:repeat)
This div should have no repeating background ()
This div should have no repeating background (background-repeat-x:no-repeat; background-repeat-y:no-repeat)
This div should have no repeating background: invalid css (background-repeat-x:repeat-y; background-repeat-y:repeat-x)
This div should have no repeating background: invalid css (background-repeat-x:repeat-x; background-repeat-y:repeat-y)
This div should have no repeating background: invalid css (background-repeat-x:10px;)
This div should have no repeating background: invalid css (background-repeat:black;)
Passed
